# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Çoklu Firma Muhasebe ve Cari Takip Paneli API
servers:
  - url: /api
    description: API temel yolu
...
paths:
...
  /gemiler:
    get:
      operationId: listGemiler
      tags: [gemiler]
      summary: Gemileri listele
      parameters:
        - name: cariId
          in: query
          schema:
            type: integer
        - name: sirketId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Gemi listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Gemi"
    post:
      operationId: createGemi
      tags: [gemiler]
      summary: Yeni gemi oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GemiInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gemi"
  /gemiler/{id}:
    get:
      operationId: getGemi
      tags: [gemiler]
      summary: Gemiyi getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gemi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GemiDetay"
    patch:
      operationId: updateGemi
      tags: [gemiler]
      summary: Gemiyi güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GemiUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Gemi"
    delete:
      operationId: deleteGemi
      tags: [gemiler]
      summary: Gemiyi sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi

  # ── Banka Hesapları ────────────────────────────────────────────────────────
...
  /faturalar:
    get:
      operationId: listFaturalar
      tags: [faturalar]
      summary: Faturaları listele
      parameters:
        - name: sirketId
          in: query
          schema:
            type: integer
        - name: cariId
          in: query
          schema:
            type: integer
        - name: durum
          in: query
          schema:
            type: string
        - name: paraBirimi
          in: query
          schema:
            type: string
        - name: baslangicTarihi
          in: query
          schema:
            type: string
            format: date
        - name: bitisTarihi
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Fatura listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fatura"
    post:
      operationId: createFatura
      tags: [faturalar]
      summary: Yeni fatura oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaturaInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fatura"
  /faturalar/{id}:
    get:
      operationId: getFatura
      tags: [faturalar]
      summary: Faturayı getir
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fatura
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FaturaDetay"
    patch:
      operationId: updateFatura
      tags: [faturalar]
      summary: Faturayı güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FaturaUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fatura"
    delete:
      operationId: deleteFatura
      tags: [faturalar]
      summary: Faturayı sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi

  # ── Ödemeler ───────────────────────────────────────────────────────────────
...
  /kdv-oranlari:
    get:
      operationId: listKdvOranlari
      tags: [kdv-oranlari]
      summary: KDV oranlarını listele
      parameters:
        - name: sirketId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: KDV oranı listesi
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/KdvOrani"
    post:
      operationId: createKdvOrani
      tags: [kdv-oranlari]
      summary: Yeni KDV oranı oluştur
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KdvOraniInput"
      responses:
        "201":
          description: Oluşturuldu
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KdvOrani"
  /kdv-oranlari/{id}:
    patch:
      operationId: updateKdvOrani
      tags: [kdv-oranlari]
      summary: KDV oranını güncelle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KdvOraniUpdate"
      responses:
        "200":
          description: Güncellendi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KdvOrani"
    delete:
      operationId: deleteKdvOrani
      tags: [kdv-oranlari]
      summary: KDV oranını sil
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Silindi

  # ── Fatura Serileri ────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Gemi:
      type: object
      required: [id, cariId, ad, olusturmaTarihi]
      properties:
        id:
          type: integer
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        sirketId:
          type: ["integer", "null"]
        ad:
          type: string
        imoNumarasi:
          type: ["string", "null"]
        bayrakDevleti:
          type: ["string", "null"]
        notlar:
          type: ["string", "null"]
        aktif:
          type: boolean
        aktifPlan:
          type: ["string", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
    GemiDetay:
      type: object
      required: [id, cariId, ad, olusturmaTarihi]
      properties:
        id:
          type: integer
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        ad:
          type: string
        imoNumarasi:
          type: ["string", "null"]
        bayrakDevleti:
          type: ["string", "null"]
        notlar:
          type: ["string", "null"]
        aktif:
          type: boolean
        starlinkPlanlari:
          type: array
          items:
            $ref: "#/components/schemas/StarlinkPlani"
        ekipmanlar:
          type: array
          items:
            $ref: "#/components/schemas/Ekipman"
        faturalar:
          type: array
          items:
            $ref: "#/components/schemas/Fatura"
        olusturmaTarihi:
          type: string
          format: date-time
    GemiInput:
      type: object
      required: [cariId, ad]
      properties:
        cariId:
          type: integer
        ad:
          type: string
        imoNumarasi:
          type: string
        bayrakDevleti:
          type: string
        notlar:
          type: string
        aktif:
          type: boolean
    GemiUpdate:
      type: object
      properties:
        ad:
          type: string
        imoNumarasi:
          type: string
        bayrakDevleti:
          type: string
        notlar:
          type: string
        aktif:
          type: boolean
        cariId:
          type: integer

    # ── Banka Hesabı ──────────────────────────────────────────────────────
...
    Fatura:
      type: object
      required: [id, sirketId, cariId, faturaNo, faturaTarihi, vadeTarihi, paraBirimi, durum, toplamTutar, kdvTutari, genelToplam, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: ["integer", "null"]
        gemiAd:
          type: ["string", "null"]
        faturaNo:
          type: string
        faturaTarihi:
          type: string
          format: date
        vadeTarihi:
          type: string
          format: date
        paraBirimi:
          type: string
        durum:
          type: string
          enum: [acik, kismi_odendi, odendi, iptal]
        toplamTutar:
          type: number
        kdvTutari:
          type: number
        genelToplam:
          type: number
        odenenTutar:
          type: number
        kalanTutar:
          type: number
        notlar:
          type: ["string", "null"]
        aciklama:
          type: ["string", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
    FaturaDetay:
      type: object
      required: [id, sirketId, cariId, faturaNo, faturaTarihi, vadeTarihi, paraBirimi, durum, toplamTutar, kdvTutari, genelToplam, olusturmaTarihi, kalemler]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: ["integer", "null"]
        gemiAd:
          type: ["string", "null"]
        faturaNo:
          type: string
        faturaTarihi:
          type: string
          format: date
        vadeTarihi:
          type: string
          format: date
        paraBirimi:
          type: string
        durum:
          type: string
        toplamTutar:
          type: number
        kdvTutari:
          type: number
        genelToplam:
          type: number
        odenenTutar:
          type: number
        kalanTutar:
          type: number
        notlar:
          type: ["string", "null"]
        aciklama:
          type: ["string", "null"]
        kalemler:
          type: array
          items:
            $ref: "#/components/schemas/FaturaKalemi"
        odemeler:
          type: array
          items:
            $ref: "#/components/schemas/Odeme"
        olusturmaTarihi:
          type: string
          format: date-time
    FaturaKalemi:
      type: object
      required: [id, faturaId, aciklama, miktar, birimFiyat, kdvOrani, araToplam, kdvTutari, genelToplam]
      properties:
        id:
          type: integer
        faturaId:
          type: integer
        aciklama:
          type: string
        miktar:
          type: number
        birimFiyat:
          type: number
        kdvOrani:
          type: number
        araToplam:
          type: number
        kdvTutari:
          type: number
        genelToplam:
          type: number
    FaturaInput:
      type: object
      required: [sirketId, cariId, faturaTarihi, vadeTarihi, paraBirimi, kalemler]
      properties:
        sirketId:
          type: integer
        cariId:
          type: integer
        gemiId:
          type: ["integer", "null"]
        faturaSerisiId:
          type: ["integer", "null"]
        faturaTarihi:
          type: string
          format: date
        vadeTarihi:
          type: string
          format: date
        paraBirimi:
          type: string
        notlar:
          type: string
        aciklama:
          type: string
        kalemler:
          type: array
          items:
            $ref: "#/components/schemas/FaturaKalemiInput"
    FaturaKalemiInput:
      type: object
      required: [aciklama, miktar, birimFiyat, kdvOrani]
      properties:
        aciklama:
          type: string
        miktar:
          type: number
        birimFiyat:
          type: number
        kdvOrani:
          type: number
    FaturaUpdate:
      type: object
      properties:
        vadeTarihi:
          type: string
          format: date
        notlar:
          type: string
        aciklama:
          type: string
        durum:
          type: string

    # ── Ödeme ────────────────────────────────────────────────────────────
    Odeme:
      type: object
      required: [id, sirketId, cariId, tip, tarih, tutar, paraBirimi, odemeYontemi, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: ["integer", "null"]
        gemiAd:
          type: ["string", "null"]
        bankaHesabiId:
          type: ["integer", "null"]
        bankaHesabiAd:
          type: ["string", "null"]
        faturaId:
          type: ["integer", "null"]
        faturaNo:
          type: ["string", "null"]
        tip:
          type: string
          enum: [tahsilat, odeme]
        tarih:
          type: string
          format: date
        tutar:
          type: number
        paraBirimi:
          type: string
        odemeYontemi:
          type: string
          enum: [banka_havalesi, eft, nakit, kredi_karti, wise, paypal, diger]
        aciklama:
          type: ["string", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
...
    StarlinkPlani:
      type: object
      required: [id, sirketId, cariId, gemiId, planAdi, baslangicTarihi, bitisTarihi, aylikUcret, paraBirimi, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        sirketAd:
          type: ["string", "null"]
        cariId:
          type: integer
        cariAd:
          type: ["string", "null"]
        gemiId:
          type: integer
        gemiAd:
          type: ["string", "null"]
        planAdi:
          type: string
        hizMbps:
          type: ["integer", "null"]
        baslangicTarihi:
          type: string
          format: date
        bitisTarihi:
          type: string
          format: date
        aylikUcret:
          type: number
        paraBirimi:
          type: string
        otomatikYenileme:
          type: boolean
        aktif:
          type: boolean
        notlar:
          type: ["string", "null"]
        kalanGun:
          type: ["integer", "null"]
        olusturmaTarihi:
          type: string
          format: date-time
...
    Ekipman:
      type: object
      required: [id, sirketId, gemiId, tip, seriNo, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        gemiId:
          type: integer
        gemiAd:
          type: ["string", "null"]
        tip:
          type: string
        seriNo:
          type: string
        kurulumTarihi:
          type: ["string", "null"]
        garantiBitisTarihi:
          type: ["string", "null"]
        notlar:
          type: ["string", "null"]
        aktif:
          type: boolean
        olusturmaTarihi:
          type: string
          format: date-time
...
    KdvOrani:
      type: object
      required: [id, sirketId, ad, oran, olusturmaTarihi]
      properties:
        id:
          type: integer
        sirketId:
          type: integer
        ad:
          type: string
        oran:
          type: number
        varsayilan:
          type: boolean
        olusturmaTarihi:
          type: string
          format: date-time
    KdvOraniInput:
      type: object
      required: [sirketId, ad, oran]
      properties:
        sirketId:
          type: integer
        ad:
          type: string
        oran:
          type: number
        varsayilan:
          type: boolean
    KdvOraniUpdate:
      type: object
      properties:
        ad:
          type: string
        oran:
          type: number
        varsayilan:
          type: boolean

    # ── Fatura Serisi ─────────────────────────────────────────────────────
...