# lib/api-zod/src/generated/types/gemiDetay.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Çoklu Firma Muhasebe ve Cari Takip Paneli API
 * OpenAPI spec version: 0.1.0
 */
import type { Ekipman } from './ekipman';
import type { Fatura } from './fatura';
import type { StarlinkPlani } from './starlinkPlani';

export interface GemiDetay {
  id: number;
  cariId: number;
  /** @nullable */
  cariAd?: string | null;
  ad: string;
  /** @nullable */
  imoNumarasi?: string | null;
  /** @nullable */
  bayrakDevleti?: string | null;
  /** @nullable */
  notlar?: string | null;
  aktif?: boolean;
  starlinkPlanlari?: StarlinkPlani[];
  ekipmanlar?: Ekipman[];
  faturalar?: Fatura[];
  olusturmaTarihi: Date;
}