# lib/api-zod/src/generated/types/kdvOrani.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Çoklu Firma Muhasebe ve Cari Takip Paneli API
 * OpenAPI spec version: 0.1.0
 */

export interface KdvOrani {
  id: number;
  sirketId: number;
  ad: string;
  oran: number;
  varsayilan?: boolean;
  olusturmaTarihi: Date;
}